Add the feature of redis persistence.Modified the relevant configuration.
