+++ build.gradle
@@ -56,6 +56,7 @@ repositories {
         maven { url "https://dl.bintray.com/ethereum/maven/" }
         maven { url "https://oss.sonatype.org/content/repositories/snapshots" }
         maven { url "https://oss.sonatype.org/content/groups/public/" }
+        maven { url 'http://maven.aliyun.com/nexus/content/groups/public/' }
     } else {
         maven {
             url 'dependencies'
@@ -68,15 +69,18 @@ List lombok = [
 ]
 
 List logger = [
-        "org.slf4j:jul-to-slf4j:1.7.25",
-        "org.apache.logging.log4j:log4j-api:2.9.0",
-        "org.apache.logging.log4j:log4j-core:2.9.0",
-        "org.apache.logging.log4j:log4j-slf4j-impl:2.9.0",
-        "org.apache.logging.log4j:log4j-web:2.9.0",
-        "org.apache.logging.log4j:log4j-jcl:2.9.0",
+        "org.slf4j:jul-to-slf4j:1.7.30",
+        "org.apache.logging.log4j:log4j-api:2.13.3",
+        "org.apache.logging.log4j:log4j-core:2.13.3",
+        "org.apache.logging.log4j:log4j-slf4j-impl:2.13.3",
+        "org.apache.logging.log4j:log4j-web:2.13.3",
+        "org.apache.logging.log4j:log4j-jcl:2.13.3",
         "com.lmax:disruptor:3.3.7"
 ]
 
+List fastjson = [
+        "com.alibaba:fastjson:1.2.58"
+]
 // junit test
 List junit = [
         "junit:junit:4.12"
@@ -93,14 +97,19 @@ List jmockit = [
 ]
 
 List json = [
-        "com.fasterxml.jackson.core:jackson-databind:2.10.0",
+        "com.fasterxml.jackson.core:jackson-databind:2.11.0",
+        "com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.11.0",
         "com.github.fge:json-schema-validator:2.2.6",
         "com.github.reinert:jjschema:1.16",
         "org.jsonschema2pojo:jsonschema2pojo-core:1.0.1"
 ]
 
+List redisson = [
+        "org.redisson:redisson:3.13.1"
+]
+
 List mysql_driver = [
-        "mysql:mysql-connector-java:8.0.16",
+        "mysql:mysql-connector-java:8.0.20",
         "org.apache.commons:commons-dbcp2:2.5.0"
 ]
 
@@ -139,14 +148,17 @@ dependencies {
     localDeps 'org.projectlombok:lombok:1.18.10'
     if (gradleVer.startsWith("4")) {
         if (!gradle.startParameter.isOffline()) {
-            compile logger, lombok, apache_commons, json, mysql_driver, zxing, rpc, pdfbox, protobuf, caffeine, oval
-            compile("com.webank:weid-contract-java:1.2.20") {
-                exclude group: "org.fisco-bcos", module: "web3sdk"
+            compile logger, lombok, apache_commons, json, fastjson, mysql_driver, redisson, zxing, rpc, pdfbox, protobuf, caffeine, oval
+            compile("com.webank:weid-contract-java:1.2.21") {
                 exclude group: "org.slf4j", module: "slf4j-log4j12"
+                // exclude group: "org.fisco-bcos", module: "web3sdk"
+                // exclude group: "org.fisco-bcos", module: "web3sdk-weevent"
             }
-            compile("org.fisco-bcos:web3sdk:2.4.0")
+            // Uncomment these two lines and exclude those above to specify web3sdk versions
+            // compile("org.fisco-bcos:web3sdk:2.4.1")
+            // compile("org.fisco-bcos:web3sdk-weevent:1.2.6")
             compile fileTree(dir: 'lib', include: '*.jar')
-            testCompile logger, lombok, apache_commons, json, junit, jmockit, rpc, pdfbox, protobuf, caffeine, oval
+            testCompile logger, lombok, apache_commons, json, fastjson, junit, jmockit, rpc, pdfbox, protobuf, caffeine, oval
         } else {
             compile fileTree(dir: 'dist/lib', include: '*.jar')
         }
@@ -157,12 +169,15 @@ dependencies {
             annotationProcessor 'org.projectlombok:lombok:1.18.10'
             testAnnotationProcessor 'org.projectlombok:lombok:1.18.10'
             testCompileOnly 'org.projectlombok:lombok:1.18.10'
-            compile logger, apache_commons, json, mysql_driver, zxing, rpc, pdfbox, protobuf, caffeine, oval
-            compile("com.webank:weid-contract-java:1.2.20") {
-                exclude group: "org.fisco-bcos", module: "web3sdk"
+            compile logger, apache_commons, json, fastjson, mysql_driver, redisson, zxing, rpc, pdfbox, protobuf, caffeine, oval
+            compile("com.webank:weid-contract-java:1.2.21") {
                 exclude group: "org.slf4j", module: "slf4j-log4j12"
+                // exclude group: "org.fisco-bcos", module: "web3sdk"
+                // exclude group: "org.fisco-bcos", module: "web3sdk-weevent"
             }
-            compile("org.fisco-bcos:web3sdk:2.4.0")
+            // Uncomment these two lines and exclude those above to specify web3sdk versions
+            // compile("org.fisco-bcos:web3sdk:2.4.1")
+            // compile("org.fisco-bcos:web3sdk-weevent:1.2.6")
 			compile fileTree(dir: 'lib', include: '*.jar')
             testCompile logger, apache_commons, json, junit, jmockit, rpc, pdfbox, protobuf, caffeine, oval
         } else {
