+++ src/main/java/com/webank/weid/util/BatchTransactionUtils.java
@@ -13,14 +13,14 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.webank.weid.constant.DataDriverConstant;
+import com.webank.weid.constant.MysqlDriverConstant;
 import com.webank.weid.constant.ErrorCode;
 import com.webank.weid.constant.ParamKeyConstant;
 import com.webank.weid.constant.WeIdConstant;
 import com.webank.weid.protocol.request.TransactionArgs;
 import com.webank.weid.protocol.response.ResponseData;
-import com.webank.weid.suite.api.persistence.Persistence;
-import com.webank.weid.suite.persistence.sql.driver.MysqlDriver;
+import com.webank.weid.suite.api.persistence.inf.Persistence;
+import com.webank.weid.suite.persistence.mysql.driver.MysqlDriver;
 
 /**
  * 批量交易处理类.
@@ -98,12 +98,12 @@ public class BatchTransactionUtils {
         if (!StringUtils.isBlank(secretKey)) {
             return secretKey;
         } else {
-            ResponseData<String> dbResp = getDataDriver().get(DataDriverConstant.DOMAIN_ENCRYPTKEY,
+            ResponseData<String> dbResp = getDataDriver().get(MysqlDriverConstant.DOMAIN_ENCRYPTKEY,
                 PropertyUtils.getProperty("blockchain.orgid"));
             Integer errorCode = dbResp.getErrorCode();
             if (errorCode != ErrorCode.SUCCESS.getCode()) {
                 logger
-                    .error("[writeTransaction] save encrypt private key to db failed.errorcode:{}",
+                    .error("[writeTransaction] add encrypt private key to db failed.errorcode:{}",
                         errorCode);
                 return null;
             }
@@ -228,7 +228,7 @@ public class BatchTransactionUtils {
         //    .saveOrUpdate(DataDriverConstant.DOMAIN_ENCRYPTKEY, weId, encryptKey);
         //Integer errorCode = dbResp.getErrorCode();
         //if (errorCode != ErrorCode.SUCCESS.getCode()) {
-        //    logger.error("[writeTransaction] save encrypt private key to db failed.errorcode:{}",
+        //    logger.error("[writeTransaction] add encrypt private key to db failed.errorcode:{}",
         //        errorCode);
         //    return null;
         //}
